backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart>=0.0.7
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0
langchain-openai
langchain
